organize tests in modules
