+++ test-framework/sudo-compliance-tests/src/macros.rs
@@ -1,0 +1,11 @@
+macro_rules! assert_contains {
+    ($haystack:expr, $needle:expr) => {
+        let haystack = &$haystack;
+        let needle = &$needle;
+
+        assert!(
+            haystack.contains(needle),
+            "{haystack:?} did not contain {needle:?}"
+        )
+    };
+}
